feat: GraphQL Apollo Client Implementation
