+++ app/src/main/java/com/example/cleanarchtest/CATApplication.kt
@@ -1,7 +1,10 @@
 package com.example.cleanarchtest
 
 import android.app.Application
-import com.example.cleanarchtest.di.*
+import com.example.cleanarchtest.di.NetworkModule
+import com.example.cleanarchtest.di.RepositoryModule
+import com.example.cleanarchtest.di.UseCaseModule
+import com.example.cleanarchtest.di.ViewModelModule
 import org.koin.android.ext.koin.androidLogger
 import org.koin.core.context.startKoin
 import org.koin.core.logger.Level
@@ -14,7 +17,6 @@ class CATApplication : Application() {
         startKoin {
             androidLogger(level = Level.INFO)
             modules(
-                ApolloNetworkModule,
                 NetworkModule,
                 UseCaseModule,
                 ViewModelModule,
