+++ app/src/main/java/com/example/cleanarchtest/di/RepositoryModule.kt
@@ -1,12 +1,12 @@
 package com.example.cleanarchtest.di
 
-import com.example.data.repository.ApolloCountryRepositoryImpl
+import com.example.data.repository.CountryRepositoryImpl
 import com.example.data.repository.PostRepositoryImpl
-import com.example.domain.repository.ApolloCountryRepository
+import com.example.domain.repository.CountryRepository
 import com.example.domain.repository.PostRepository
 import org.koin.dsl.module
 
 val RepositoryModule = module {
-    single<PostRepository> { PostRepositoryImpl(get()) }
-    single<ApolloCountryRepository> { ApolloCountryRepositoryImpl(get()) }
+    single<PostRepository> { PostRepositoryImpl(postService = get()) }
+    single<CountryRepository> { CountryRepositoryImpl(apolloClient = get()) }
 }
