+++ app/src/main/java/com/example/cleanarchtest/di/UseCaseModule.kt
@@ -1,13 +1,12 @@
 package com.example.cleanarchtest.di
 
 import com.example.domain.model.PostUseCases
-import com.example.domain.useCase.ApolloGetCountryUseCase
+import com.example.domain.useCase.GetCountryUseCase
 import com.example.domain.useCase.GetAllPostUseCase
 import org.koin.dsl.module
 
 val UseCaseModule = module {
     single { GetAllPostUseCase(postRepository = get()) }
     single { PostUseCases(getAllPostUseCase = get()) }
-
-    single { ApolloGetCountryUseCase(apolloCountryRepository = get()) }
+    single { GetCountryUseCase(countryRepository = get()) }
 }
