+++ app/src/main/java/com/example/cleanarchtest/di/ViewModelModule.kt
@@ -5,5 +5,10 @@ import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
 val ViewModelModule = module {
-    viewModel { PostListVM(postUseCases = get(), countryUseCase = get()) }
+    viewModel {
+        PostListVM(
+            postUseCases = get(),
+            countryUseCase = get()
+        )
+    }
 }
