+++ app/src/main/java/com/example/cleanarchtest/ui/postList/PostListVM.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.viewModelScope
 import com.example.cleanarchtest.util.UiState
 import com.example.domain.model.Post
 import com.example.domain.model.PostUseCases
-import com.example.domain.useCase.ApolloGetCountryUseCase
+import com.example.domain.useCase.GetCountryUseCase
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -14,7 +14,7 @@ import kotlinx.coroutines.launch
 
 class PostListVM(
     private val postUseCases: PostUseCases,
-    private val countryUseCase: ApolloGetCountryUseCase,
+    private val countryUseCase: GetCountryUseCase,
     private val dispatcher: CoroutineDispatcher = Dispatchers.IO
 ) : ViewModel() {
 
@@ -28,7 +28,6 @@ class PostListVM(
     private fun fetchAllPostFromNetwork() {
         viewModelScope.launch(dispatcher) {
             _uiState.value = UiState.Loading
-//            val postList = postUseCases.getAllPostUseCase()
             val postList = countryUseCase()
             _uiState.value = UiState.Success(data = postList)
         }
