--- domain/src/main/java/com/example/domain/useCase/ApolloGetCountryUseCase.kt
@@ -1,7 +1,0 @@
-package com.example.domain.useCase
-
-import com.example.domain.repository.ApolloCountryRepository
-
-class ApolloGetCountryUseCase(private val apolloCountryRepository: ApolloCountryRepository) {
-    suspend operator fun invoke() = apolloCountryRepository.getCountries()
-}
