+++ domain/src/main/java/com/example/domain/useCase/GetCountryUseCase.kt
@@ -1,0 +1,7 @@
+package com.example.domain.useCase
+
+import com.example.domain.repository.CountryRepository
+
+class GetCountryUseCase(private val countryRepository: CountryRepository) {
+    suspend operator fun invoke() = countryRepository.getCountries()
+}
